Add DRYRUN option to deploy (#19)
* Add DRYRUN option to deploy

* Add dryrun deploy to test suite

* Remove newline

+++ .circleci/deploy-ghpages.sh
@@ -46,8 +46,11 @@ mv ../_site/* ./
 git add -A
 # now commit, ignoring branch gh-pages doesn't seem to work, so trying skip
 git commit --allow-empty -m "Deploy to GitHub pages [ci skip]"
-# and push, but send any output to /dev/null to hide anything sensitive
-git push --force --quiet origin gh-pages
+  
+if [ -z "$DRYRUN" ]; then
+  # and push, but send any output to /dev/null to hide anything sensitive
+  git push --force --quiet origin gh-pages
+fi
 # go back to where we started and remove the gh-pages git repo we made and used
 # for deployment
 cd ..
